+++ src/com/wingfac/MaitreyaRim/po/OrderInformation.java
@@ -18,11 +18,11 @@ public class OrderInformation implements Serializable {
 	private String payment_time;
 	private Double o_amount;
 	private Double amount_paid;
-	private Integer use_limit_integral;
-	private Integer use_permanent_points;
+	private Double use_limit_integral;
+	private Double use_permanent_points;
 	private Double ermanent_integral_bonus;
 	private Double time_limited_integration;
-	private Integer full_integral_purchase;
+	private Double full_integral_purchase;
 	//支付中:2 新加的支付状态 
 	private String o_state;
 	private String store_consume_state;
@@ -38,9 +38,9 @@ public class OrderInformation implements Serializable {
 			String o_number, String s_name, String s_mobile,
 			String auBuyerNick, String auMobile, String auAddress,
 			String o_time, String payment_time, Double o_amount,
-			Double amount_paid, Integer use_limit_integral,
-			Integer use_permanent_points, Double ermanent_integral_bonus,
-			Double time_limited_integration, Integer full_integral_purchase,
+			Double amount_paid, Double use_limit_integral,
+			Double use_permanent_points, Double ermanent_integral_bonus,
+			Double time_limited_integration, Double full_integral_purchase,
 			String o_state, String store_consume_state, Integer pay_state) {
 		super();
 		this.o_id = o_id;
@@ -187,19 +187,19 @@ public class OrderInformation implements Serializable {
 		this.amount_paid = amount_paid;
 	}
 
-	public Integer getUse_limit_integral() {
+	public Double getUse_limit_integral() {
 		return use_limit_integral;
 	}
 
-	public void setUse_limit_integral(Integer use_limit_integral) {
+	public void setUse_limit_integral(Double use_limit_integral) {
 		this.use_limit_integral = use_limit_integral;
 	}
 
-	public Integer getUse_permanent_points() {
+	public Double getUse_permanent_points() {
 		return use_permanent_points;
 	}
 
-	public void setUse_permanent_points(Integer use_permanent_points) {
+	public void setUse_permanent_points(Double use_permanent_points) {
 		this.use_permanent_points = use_permanent_points;
 	}
 
@@ -219,11 +219,11 @@ public class OrderInformation implements Serializable {
 		this.time_limited_integration = time_limited_integration;
 	}
 
-	public Integer getFull_integral_purchase() {
+	public Double getFull_integral_purchase() {
 		return full_integral_purchase;
 	}
 
-	public void setFull_integral_purchase(Integer full_integral_purchase) {
+	public void setFull_integral_purchase(Double full_integral_purchase) {
 		this.full_integral_purchase = full_integral_purchase;
 	}
